Added idle/enable state for SmartPot
Added a new EEPROM address and static Uint8_t to keep track of the idle/enable state.

Added the smart_pot_set_state() method that takes and uint8_t value as to set the state of the machine.

Added an if statement for smart_pot_tryWater() to stop it from watering if the enable state is off (0).